+++ Helper/Url.php
@@ -49,6 +49,8 @@ use Nosto\Request\Http\HttpRequest;
 use Nosto\Tagging\Helper\Data as NostoDataHelper;
 use Nosto\Tagging\Model\Product\Repository as ProductRepository;
 use Nosto\Tagging\Model\Product\Url\Builder as NostoUrlBuilder;
+use Nosto\Helper\UrlHelper;
+use Nosto\Tagging\Logger\Logger as NostoLogger;
 
 /**
  * Url helper class for common URL related tasks.
@@ -136,6 +138,7 @@ class Url extends AbstractHelper
     private $backendDataHelper;
     private $productRepository;
     private $nostoUrlBuilder;
+    private $logger;
 
     /** @noinspection PhpUndefinedClassInspection */
     /**
@@ -158,7 +161,8 @@ class Url extends AbstractHelper
         UrlBuilder $urlBuilder,
         /** @noinspection PhpDeprecationInspection */
         BackendDataHelper $backendDataHelper,
-        NostoUrlBuilder $nostoUrlBuilder
+        NostoUrlBuilder $nostoUrlBuilder,
+        NostoLogger $nostoLogger
     ) {
         parent::__construct($context);
 
@@ -168,6 +172,7 @@ class Url extends AbstractHelper
         $this->nostoDataHelper = $nostoDataHelper;
         $this->backendDataHelper = $backendDataHelper;
         $this->nostoUrlBuilder = $nostoUrlBuilder;
+        $this->logger = $nostoLogger;
     }
 
     /**
@@ -276,6 +281,22 @@ class Url extends AbstractHelper
         );
         $url = $this->replaceQueryParamsInUrl(['q' => 'nosto'], $url);
         return $this->addNostoDebugParamToUrl($url);
+    }
+
+    /**
+     * Returns the store domain
+     *
+     * @param Store $store
+     * @return string
+     */
+    public function getActiveDomain(Store $store)
+    {
+        try {
+            return UrlHelper::parseDomain($store->getBaseUrl());
+        } catch (\Exception $e) {
+            $this->logger->exception($e);
+            return '';
+        }
     }
 
     /**